dim06.08.2020 22:00

+++ src/main/java/musketeers/kz/entity/standart/Reports.java
@@ -10,7 +10,7 @@ public class Reports {
     private String video;
     private String location;
     private int isDone;
-    private int chatID;
+    private Long chatID;
     private String fullName;
     private String categoryName;
 }
